+++ src/main/java/com/ddr/model/DispatchReport.java
@@ -1,5 +1,6 @@
 package com.ddr.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class DispatchReport {
@@ -8,7 +9,7 @@ public class DispatchReport {
     private String custName;    // PARTY >> RUBY PHARMA
     private String destination; // DESTINATION >> SOLAPUR
     private String transporter; // TRANSPORTER >> BTI
-    private double goodsValue;  // GOODS VALUE >> 12934.00
+    private BigDecimal goodsValue;  // GOODS VALUE >> 12934.00
     private String invNo;   // INVOICE NO >> BWMNI2300001
     private String lrNum;   // LR NO >> 675433
     private String driverName;  // DRIVER NAME
@@ -42,8 +43,8 @@ public class DispatchReport {
 //    private int free;
 //    private int repl;
 //    private float rate;
-//    private double tax;
-//    private double discount;
+//    private BigDecimal tax;
+//    private BigDecimal discount;
 
     public DispatchReport() {
     }
@@ -88,11 +89,11 @@ public class DispatchReport {
         this.transporter = transporter;
     }
 
-    public double getGoodsValue() {
+    public BigDecimal getGoodsValue() {
         return goodsValue;
     }
 
-    public void setGoodsValue(double goodsValue) {
+    public void setGoodsValue(BigDecimal goodsValue) {
         this.goodsValue = goodsValue;
     }
